This PR:

- Adds Permission based Authorization. The goal is to be able to assign multiple permissions to any user but I hit what is looks like a Typeorm bug regarding Enum Arrays. As I was spending more than an hour on it, I decided to go with only on permission per User to avoid this bug, it is now the same thing as a Role Based Authorization scheme. =( 
- Adds queries and mutations that uses that Authorization.

```
=============================== Coverage summary ===============================
Statements   : 72.91% ( 603/827 )
Branches     : 76.19% ( 32/42 )
Functions    : 26.73% ( 27/101 )
Lines        : 72.91% ( 603/827 )
================================================================================
```

+++ src/task/task.service.ts
@@ -3,16 +3,25 @@ import { Task } from './entities/task.entity';
 import { CreateTaskInput } from './dto/create-task.input';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { User } from '../users/entities/user.entity';
 
 @Injectable()
 export class TaskService {
   constructor(
     @InjectRepository(Task)
     private taskRepository: Repository<Task>,
+    @InjectRepository(User)
+    private userRepository: Repository<User>,
   ) {}
 
-  async create(createTaskInput: CreateTaskInput) {
-    const result = this.taskRepository.create(createTaskInput);
+  async create(createTaskInput: CreateTaskInput, userId: string) {
+    if (userId === null) {
+      const result = await this.taskRepository.create(createTaskInput);
+      return this.taskRepository.save(result);
+    }
+
+    const user: User = await this.userRepository.findOneBy({ id: userId });
+    const result = this.taskRepository.create({ ...createTaskInput, user });
     return this.taskRepository.save(result);
   }
 
@@ -20,17 +29,67 @@ export class TaskService {
     return this.taskRepository.find();
   }
 
-  async findOne(id: string): Promise<Task> {
+  async findByUserId(userId: string): Promise<Task[]> {
+    const user: User = await this.userRepository.findOneBy({ id: userId });
+    const result = await this.taskRepository.findBy({ user });
+    const taskWithUser = [];
+    result.forEach((task) =>
+      taskWithUser.push({
+        title: task.title,
+        status: task.status,
+        id: task.id,
+        description: task.description,
+        user,
+      }),
+    );
+    return taskWithUser;
+  }
+
+  async findOne(id: string, userId: string = null): Promise<Task> {
+    if (userId === null) return this.taskRepository.findOneBy({ id });
+
+    const user: User = await this.userRepository.findOneBy({ id: userId });
+    const result = await this.taskRepository.findOneBy({ id, user });
+
+    return {
+      title: result.title,
+      status: result.status,
+      id: result.id,
+      description: result.description,
+      user,
+    };
+  }
+
+  async update(
+    id: string,
+    updateTaskInput: Partial<Task>,
+    userId: string,
+  ): Promise<Task> {
+    if (userId === null) {
+      await this.taskRepository.update(id, updateTaskInput);
+    } else {
+      const user: User = await this.userRepository.findOneBy({ id: userId });
+      await this.taskRepository.update(id, { ...updateTaskInput, user });
+    }
     return this.taskRepository.findOneBy({ id });
   }
 
-  async update(id: string, updateTaskInput: Partial<Task>): Promise<Task> {
-    await this.taskRepository.update(id, updateTaskInput);
-    return this.taskRepository.findOneBy({ id });
-  }
+  async remove(id: string, userId: string): Promise<boolean> {
+    let result;
 
-  async remove(id: string): Promise<boolean> {
-    const result = await this.taskRepository.delete(id);
-    return result.affected === 1;
+    if (userId === null) {
+      result = await this.taskRepository.delete(id);
+    } else {
+      const user: User = await this.userRepository.findOneBy({ id: userId });
+      const tasks: Task[] = await this.taskRepository.findBy({
+        id,
+        user: user,
+      });
+      if (tasks.length === 1) {
+        result = await this.taskRepository.delete(id);
+      }
+    }
+
+    return result?.affected === 1;
   }
 }
